moved image component to script folder
