+++ blocks/category-family/category-family.js
@@ -1,5 +1,5 @@
 /* eslint-disable import/no-unresolved */
-import { getMetadata } from '../../scripts/lib-franklin.js';
+import { getMetadata, loadScript } from '../../scripts/lib-franklin.js';
 import { getCookie, isOTEnabled } from '../../scripts/scripts.js';
 
 const categoryFamily = `
@@ -159,6 +159,7 @@ export default async function decorate(block) {
   block.innerHTML = categoryFamily;
   await import('https://static.cloud.coveo.com/atomic/v2/atomic.esm.js');
   await customElements.whenDefined('atomic-search-interface');
+  loadScript('../../scripts/image-component.js');
 
   const categorySearchInterface = document.querySelector('atomic-search-interface.category-search');
