+++ blocks/product-resources/product-resources.js
@@ -1,4 +1,5 @@
 /* eslint-disable import/no-unresolved */
+import { loadScript } from '../../scripts/lib-franklin.js';
 import {
   getCookie, isOTEnabled, getProductResponse, getSKU,
 } from '../../scripts/scripts.js';
@@ -165,6 +166,7 @@ export default async function decorate(block) {
     block.innerHTML = productResources;
     await import('https://static.cloud.coveo.com/atomic/v2/atomic.esm.js');
     await customElements.whenDefined('atomic-search-interface');
+    loadScript('../../scripts/image-component.js');
 
     const resourceSearchInterface = document.querySelector('atomic-search-interface.resource-search');
